Update calculator.py
changed int for float

+++ calculator.py
@@ -1,4 +1,4 @@
-x = int (input ("What's X? "))
-y = int (input ("What's Y? "))
+x = float (input ("What's X? "))
+y = float (input ("What's Y? "))
 
 print (x+y)
